hypergeometric test moved from resources/ to common/
